feat: adds base flask web server and questions, hospital, and user tables.
This commit adds the base flask webserver in src/python/server.
Additionally, it creates the migrations folder for use with
Flask-Migrate and the Questions, Hospitals, and User database
tables. These tables use SQLAlchemy. There is also a simple
README.md file in the root directory which was created to help
others getting started.