feat: Maintain selected options for each questions in a list
This commit adds the responses list to the question_card.dart file.
When a user selects an option, it adds a dictionary to the responses
list. This will eventually be sent to the server.

This commit also adds the collection dependency from dart's
pub.dev to the pubspec.yml. This is used for comparing dictionaries
because dart doesn't inheritly know how to compare two dictionaries
for equality.