feat: Add QR code scanner
This commit adds the barcode_scan dependency and the barcode scan
screen. There are a lot of small changes to files under the ios/
directory. These include minor changes to the Podfile as well
as some extra files. All of this is to handle dependencies and
other issues that came up when implementing the QR code scanner.